Introduce ProjectModelProvider
This is mainly for RLS which would love to be a single source of truth
for project info, and which would rather link to a single copy of
Cargo.

However, it might also be possible to use other things as sources of
project information, for example, `cargo metadata` or Bazel/Buck build
rules.
