+++ Cargo.toml
@@ -15,13 +15,14 @@ path = "src/racer/lib.rs"
 name = "racer"
 path = "src/bin/main.rs"
 doc = false
+required-features = ["cargo"]
 
 [profile.release]
 debug = true
 
 [dependencies]
 bitflags = "1.0"
-cargo = "0.28"
+cargo = { version = "0.28", optional = true }
 log = "0.4"
 rustc-ap-syntax = "209.0.0"
 env_logger = "0.5"
@@ -37,6 +38,7 @@ path = "testutils"
 
 
 [features]
+default = ["cargo"]
 nightly = []
 
 [workspace]
